Split root view and rootviewmodel into separate files
